% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomerative_cover.R
\name{agglomerative_cover}
\alias{agglomerative_cover}
\title{Agglomerative cover}
\usage{
agglomerative_cover(cover)
}
\arguments{
\item{cover}{original cover}
}
\description{
Get agglomerative cover
}
\examples{
# generate data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rbind(rcircle(200, 1, .1), 
                     data.frame(x = rnorm(100, 0, .1), 
                                y = rnorm(100, 0, .1)))

# run divisive cover
dc <- divisive_cover(distance_matrix = dist(data_matrix), 
                     relative_diameter = 0.2, relative_distance = 0.3)
# get subcover
sc <- subcover(dc, 0.3, "snapshot")

# agglomerative cover from subcover
ac <- agglomerative_cover(sc)
}

