% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R, R/cover_methods.R
\docType{class}
\name{cover-class}
\alias{cover-class}
\alias{patch-class}
\alias{show,cover-method}
\alias{show,patch-method}
\title{An S4 class to represent a cover.}
\usage{
\S4method{show}{cover}(object)

\S4method{show}{patch}(object)
}
\arguments{
\item{object}{A cover}
}
\description{
An S4 class to represent a cover.

An S4 class to represent a patch.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame}

\item{\code{subsets}}{A list of patches}

\item{\code{internal_nodes}}{internal nodes}

\item{\code{external_nodes}}{external nodes}

\item{\code{data_filter_values}}{the filter value of the entire data set}

\item{\code{parameter}}{list of parameters}

\item{\code{type}}{type of cover}

\item{\code{id}}{Patch id}

\item{\code{indices}}{Indices of data points in patch}

\item{\code{predicted}}{Indices of predicted points}

\item{\code{anchor_points}}{Anchor points}

\item{\code{filter_value}}{Filter value}

\item{\code{parent_filter}}{Filter value of parent}
}}

