% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_covers.R
\name{cover_distance}
\alias{cover_distance}
\title{Distance between covers}
\usage{
cover_distance(..., metric = c("vi", "bottleneck", "interleaving"), dim = 1)
}
\arguments{
\item{...}{covers to compare}

\item{metric}{which metric should be calculated}

\item{dim}{dimension(s) for bottleneck distance}
}
\description{
Compare two covers using persistent homology or entropy
}
\examples{
# generate data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)
# calculate covers
dc1 <- divisive_cover(data = data_matrix,
                      stop_fct = stop_relative_filter(relative_filter = 0.1))
dc2 <- divisive_cover(data = data_matrix, 
                      stop_fct = stop_relative_filter(relative_filter = 0.3))
cover1 <- subcover(dc1, relative_filter = 0.1, method = "snapshot")
cover2 <- subcover(dc2, relative_filter = 0.5, method = "snapshot")

cover_distance(cover1, cover2, metric = "vi")
cover_distance(cover1, cover2, metric = "bottleneck", dim = 0)
cover_distance(cover1, cover2, metric = "bottleneck", dim = 1)
cover_distance(cover1, cover2, metric = "interleaving")

# plot covers
\dontrun{
plot(cover1)
plot(cover2)
plot(c(cover1, cover2))
}
}
