% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_matrix.R
\name{cover_matrix}
\alias{cover_matrix}
\alias{predict_cover_matrix}
\title{Cover matrix}
\usage{
cover_matrix(cover)

predict_cover_matrix(cover)
}
\arguments{
\item{cover}{A cover}
}
\description{
Calculate the n * m matrix C, where n is the number of data points and m is the 
number of nodes, where \deqn{C_{ij}} is 1 if the the i-th data point is in the j-th 
subset and 0 otherwise.
}

