% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter_cover.R
\name{diameter_cover}
\alias{diameter_cover}
\title{Diameter cover function}
\usage{
diameter_cover(cover, relative_diameter)
}
\arguments{
\item{cover}{A divisive cover}

\item{relative_diameter}{The diameter to extract}
}
\description{
Find the divisive cover that corresponds to a certain diameter
}
\examples{
rcircle <- function(N, r, sd){
   radius <- rnorm(N, r, sd)
   angle <- runif(N, 0, 2 * pi)
   data.frame(x = radius * cos(angle), 
              y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

dc <- divisive_cover(distance_matrix = dist(data_matrix), 
                     relative_diameter = 0.5, relative_distance = 0.2)
rc <- diameter_cover(dc, relative_diameter = 0.7)

\dontrun{
plot(data_matrix)
plot(rc)
}
}

