% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{distance_fct}
\alias{distance_fct}
\alias{distance_cdist}
\alias{distance_euclidean}
\alias{distance_matrix}
\title{Distance functions for divisive cover}
\usage{
distance_cdist(metric = "euclidean")

distance_euclidean(data, X, Y)

distance_matrix(data, X, Y)
}
\arguments{
\item{metric}{metric in `cdist`}

\item{data}{data for divisive cover}

\item{X, Y}{indices}
}
\description{
Distance functions that can be used in \code{\link{divisive_cover}}. These functions
have arguments \code{data}, \code{X} and \code{Y}, where \code{data} is the data 
matrix and \code{X} and \code{Y} are indices.
}
\details{
The function \code{distance_euclidean} returns the euclidean distance of 
\code{data[X, ]} and \code{data[Y, ]}. 

The function \code{distance_matrix} assumes that the data matrix is a distance 
matrix and returns \code{data[X, Y]}. 

The function \code{distance_cdist} takes a metric as an argument and returns a
distance function with arguments \code{data}, \code{X} and \code{Y}. See
\code{\link[rdist]{cdist}}.
}
