% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division_functions.R
\name{division_fct}
\alias{division_fct}
\alias{relative_factor_division}
\alias{relative_gap_division}
\title{Division functions}
\usage{
relative_factor_division(relative_factor)

relative_gap_division(relative_gap)
}
\arguments{
\item{relative_factor}{corresponds to (1-2 * delta)/(1+2 * delta)}

\item{relative_gap}{delta-parameter for delta-filtered cover (0 < delta <= 1/2)}
}
\description{
This function does the actual division in divisive cover.
}
\details{
The function \code{relative_factor_division} returns a function that divides
a patch into subpatches. All points with distance from one anchor point 
larger than \code{relative_factor} times the distance from the other anchor 
point come in one subpatch and vice versa. 

The function \code{relative_gap_division} uses the \code{relative_factor_division}
with \code{relative_factor = (1-2 * relative_gap)/(1+2 * relative_gap)}

Note that all division functions should return a list of two patches, where
the first patch contains the parent anchor point 1 and the second patch 
contains parent anchor point 2.
}
