% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover_classification.R
\name{divisive_cover_classification}
\alias{divisive_cover_classification}
\title{divisive cover}
\usage{
divisive_cover_classification(distance_matrix, response, delta,
  max_nodes = Inf)
}
\arguments{
\item{distance_matrix}{an n x n matrix of pairwise dissimilarities}

\item{response}{a length n response vector}

\item{delta}{delta-parameter for delta-filtered cover (0 < delta <= 1/2)}

\item{max_nodes}{maximum number of nodes the algorithm should generate}
}
\description{
This function divides the data into a cover using the distance matrix.
}
\examples{
rcircle <- function(N, r, sd){
   radius <- rnorm(N, r, sd)
   angle <- runif(N, 0, 2 * pi)
   data.frame(x = radius * cos(angle), 
              y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# calculate and update divisive cover
dc1 <- divisive_cover(distance_matrix = dist(data_matrix),
                      delta = 0.1, 
                      relative_diameter = 0.7)
dc2 <- update_divisive_cover(dc1, relative_diameter = 0.5)

# get one snapshot for plotting
ddc <- subcover(dc2, relative_diameter = 0.7, method = "snapshot")

\dontrun{
plot(data_matrix)
plot(ddc)
}
}

