% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover_classification.R
\name{divisive_cover_classify}
\alias{divisive_cover_classify}
\title{Classify}
\usage{
divisive_cover_classify(pred)
}
\arguments{
\item{pred}{a prediction cover}
}
\description{
Classify from divisive cover
}

