% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_ensemble.R
\name{divisive_prediction_ensemble}
\alias{divisive_prediction_ensemble}
\alias{predict_divisive_prediction_ensemble}
\title{Divisive prediction ensemble}
\usage{
divisive_prediction_ensemble(train, group, ntree = 100, depth = 1000,
  delta_range = c(0, 0.2))

predict_divisive_prediction_ensemble(dc_ensemble, test)
}
\arguments{
\item{train}{Training data}

\item{group}{Class of training data}

\item{ntree}{Number of trees to grow}

\item{depth}{Maximal depth of trees}

\item{delta_range}{Range of delta parameters to consider}

\item{dc_ensemble}{Output of \code{divisive_prediction_ensemble}}

\item{test}{Test data}
}
\description{
A divisive prediction ensemble is an algorithm similar to random forests with
divisive covers instead of decission trees.
}
\examples{
# data
sampled_index <- sample(150)
train <- iris[sampled_index[1:100], 1:4]
group <- iris[sampled_index[1:100], 5]
test <- iris[sampled_index[101:150], 1:4]
observed <- iris[sampled_index[101:150], 5]

# fit divisive prediction ensemble
fit <- divisive_prediction_ensemble(train, group, ntree = 10, depth = 10, delta_range = c(0, 0))
# get predictions
predictions <- predict_divisive_prediction_ensemble(fit, test)

# calculate misclassification error
table(observed, predictions)
tab <- table(observed, predictions)
misclassification_error <- 1 - sum(diag(tab)) / sum(tab)
misclassification_error
}
