% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_fct}
\alias{filter_fct}
\alias{diameter_filter}
\alias{cardinality_filter}
\alias{classification_filter}
\title{Filter functions}
\usage{
diameter_filter(patches, data, distance_fct, ...)

cardinality_filter(patches, ...)

classification_filter(patches, group, ...)
}
\arguments{
\item{patches}{Patch to apply filter}

\item{data}{Data used in distance function}

\item{distance_fct}{The used distance function}

\item{...}{ignored arguments}

\item{group}{the group for classification}
}
\description{
Filter after diameter or cardinality
}
