% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_distance.R
\name{patch_cdist}
\alias{patch_cdist}
\title{patch distances}
\usage{
patch_cdist(patches, X = 1:length(patches), Y = 1:length(patches),
  metric = "jaccard")
}
\arguments{
\item{patches}{Patches to calculate distance from}

\item{X, Y}{Indices}

\item{metric}{Used metric}
}
\description{
Calculate the distances between patches
}
\examples{
# generate sample data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# calculate divisive cover
dc <- divisive_cover(data = data_matrix,
                     division_fct = relative_gap_division(0.05), 
                     stop_fct = stop_relative_filter(0.5))

# distance of external nodes
patch_cdist(slot(subcover(dc), "subsets"))

}
