% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{persistence_from_cover}
\alias{persistence_from_cover}
\title{Persistence from cover}
\usage{
persistence_from_cover(cover)
}
\arguments{
\item{cover}{A cover}
}
\description{
Calculate the persistence from the cover
}

