% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{persistence_from_filtration}
\alias{persistence_from_filtration}
\title{Persistence from filtration}
\usage{
persistence_from_filtration(filtration)
}
\arguments{
\item{filtration}{A filtration}
}
\description{
Calculate the persistence from the filtration
}

