% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{plot.cover}
\alias{plot.cover}
\title{Plot cover output}
\usage{
\method{plot}{cover}(x, coloring = NULL, simplify = c("none", "duplicates",
  "subsets"), label = sapply(x@subsets, slot, "id"), legend = TRUE,
  device = c("plot", "tkplot"), seed = 1, ...)
}
\arguments{
\item{x}{A cover}

\item{coloring}{color values. See \code{\link{cover_coloring}}}

\item{simplify}{simplification method. See \code{\link{simplify_cover}}}

\item{label}{how should the nodes be labeled. Set NA for no labels.}

\item{legend}{should color legend be displayed}

\item{device}{use igraph::plot.igraph or igraph::tkplot}

\item{seed}{random seed used to make plots look the same}

\item{...}{additional plotting parameters. See \code{\link[igraph]{igraph.plotting}} for the complete list.}
}
\description{
\code{plot.cover} generates R plots of covers
}
\examples{
# generate sample data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# calculate divisive cover
dc <- divisive_cover(data = data_matrix,
                     division_fct = relative_gap_division(0.05), 
                     stop_fct = stop_relative_filter(0.5))

# external nodes
sc <- subcover(dc)

# distance
patches <- slot(sc, "subsets")
d_centroid <- patch_centroid_dist(data_matrix, patches, patches)

# colors 
coloring <- cover_coloring(sc, color_values = data_matrix[, 1], FUN = median)

# plot
\dontrun{
  plot(sc, coloring = coloring)
  # custom layout
  plot(sc, coloring = coloring, layout = igraph::layout_with_kk)
  plot(sc, coloring = coloring, layout = igraph::norm_coords(cmdscale(d_centroid)))
}
}
\seealso{
\code{\link{cover-class}}, \code{\link{subcover}}
}
\author{
Nello Blaser
}
\keyword{plotting}
