% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_persistence.R
\name{plot_persistence}
\alias{plot_persistence}
\title{Plot persistence}
\usage{
plot_persistence(pers, mode = c("diag", "bars"), relative = TRUE)
}
\arguments{
\item{pers}{A persistence data.frame}

\item{mode}{Either "diag" for persistence diagram or "bars" for 
barcodes}

\item{relative}{Should relative diameters be used}
}
\description{
Plot persistence diagram or barcode
}

