% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cover}
\alias{predict.cover}
\title{Predict}
\usage{
\method{predict}{cover}(object, newdists, ...)
}
\arguments{
\item{object}{a divisive cover}

\item{newdists}{a matrix of distances between the new points and the 
points used to train the divisive cover}

\item{...}{further arguments passed fromother methods}
}
\description{
Predict in which subsets new data come to lie when given a previous
divisive cover
}

