% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_cover}
\alias{predict_cover}
\alias{predict.cover}
\alias{group_from_predict_cover}
\title{Predict}
\usage{
\method{predict}{cover}(object, newdata, predict_fct, ...)

group_from_predict_cover(pc, group = NULL)
}
\arguments{
\item{object}{a divisive cover}

\item{newdata}{new data}

\item{predict_fct}{predict function (see below)}

\item{...}{further arguments passed from other methods}

\item{pc}{prediction cover}

\item{group}{specify the group of the training set (not necessary if used in divisive_cover)}
}
\description{
Predict in which subsets new data come to lie when given a previous
divisive cover

Extract group from predict cover
}
\examples{
iris <- datasets::iris

ttsplit <- sample(1:2, nrow(iris), replace = TRUE, prob = c(2, 1))
train <- iris[ttsplit == 1, ]
test <- iris[ttsplit == 2, ]

dc <- divisive_cover(data = train[, 1:4], 
                     group = train[, 5], 
                     distance_fct = distance_cdist("euclidean"), 
                     stop_fct = stop_relative_filter(relative_filter = 0), 
                     anchor_fct = anchor_classify, 
                     filter_fct = classification_filter, 
                     division_fct = relative_gap_division(relative_gap = 0.01)
)

pc <- predict(object = dc, newdata = test[, 1:4], 
              predict_fct = relative_gap_prediction(relative_gap = 0.01))
# plot
sc <- subcover(pc, 0.6, "snapshot")
col <- predict_coloring(sc)
\dontrun{
plot(sc, coloring = col)
}

predicted <- group_from_predict_cover(pc)
observed <- test[, 5]

table(observed, predicted)
}
