% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{relative_factor_prediction}
\alias{relative_factor_prediction}
\alias{relative_gap_prediction}
\title{Prediction functions}
\usage{
relative_factor_prediction(relative_factor)

relative_gap_prediction(relative_gap, euclidean = FALSE)
}
\arguments{
\item{relative_factor}{corresponds to (1-2 * delta)/(1+2 * delta)}

\item{relative_gap}{delta-parameter for delta-filtered cover (0 < delta <= 1/2)}

\item{euclidean}{logical, if TRUE then (1-delta)/(1+delta) is used instead of 
(1-2 * delta)/(1+2 * delta)}
}
\description{
The prediction function should usually take the same arguments as
the division function.
}
