% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_ensemble.R
\name{print.divisive_ensemble}
\alias{print.divisive_ensemble}
\title{Print divisive ensemble}
\usage{
\method{print}{divisive_ensemble}(x, ...)
}
\arguments{
\item{x}{divisive ensemble}

\item{...}{ignored}
}
\description{
Print divisive ensemble
}
