% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_cover.R
\name{radius_cover}
\alias{radius_cover}
\title{Radius cover function}
\usage{
radius_cover(cover, relative_radius)
}
\arguments{
\item{cover}{A divisive cover}

\item{relative_radius}{The radius to extract}
}
\description{
Find the divisive cover that corresponds to a certain radius
}
\examples{
rcircle <- function(N, r, sd){
   radius <- rnorm(N, r, sd)
   angle <- runif(N, 0, 2 * pi)
   data.frame(x = radius * cos(angle), 
              y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

dc <- divisive_cover(data_matrix, distance_object = dist, 
                     relative_radius = 0.5, relative_distance = 0.2)
rc <- radius_cover(dc, relative_radius = 0.7)

\dontrun{
plot(data_matrix)
plot(rc)
}
}

