% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division_functions.R
\name{relative_factor_division}
\alias{relative_factor_division}
\title{Division functions}
\usage{
relative_factor_division(relative_factor)
}
\arguments{
\item{relative_factor}{corresponds to (1-2 * delta)/(1+2 * delta)}

\item{relative_gap}{delta-parameter for delta-filtered cover (0 < delta <= 1/2)}
}
\description{
Division functions
}

