% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_cover.R
\name{simplify_cover}
\alias{simplify_cover}
\title{Simplify cover}
\usage{
simplify_cover(x, method = c("none", "duplicates", "subsets"))
}
\arguments{
\item{x}{cover}

\item{method}{simplification method}
}
\description{
Simplify cover by removing duplicates or subsets
}

