% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcover.R
\name{subcover}
\alias{subcover}
\title{Subcover}
\usage{
subcover(cover, relative_diameter, method = c("divisive", "snapshot",
  "range"))
}
\arguments{
\item{cover}{A divisive cover}

\item{relative_diameter}{The diameter to extract}

\item{method}{Either "divisive" to get a divisive cover or "snapshot" 
to get only the lowest level}
}
\description{
Find the (divisive) cover that corresponds to a certain diameter
}
\examples{
rcircle <- function(N, r, sd){
   radius <- rnorm(N, r, sd)
   angle <- runif(N, 0, 2 * pi)
   data.frame(x = radius * cos(angle), 
              y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

dc <- divisive_cover(distance_matrix = dist(data_matrix), 
                     delta = 0.1, 
                     stop_fct = stop_relative_diameter(relative_diameter = 0.5))
dc_0.7 <- subcover(dc, relative_diameter = 0.7, method = "divisive")
sc_0.7 <- subcover(dc, relative_diameter = 0.7, method = "snapshot")

\dontrun{
plot(data_matrix)
plot(sc_0.7)
}
}

