% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_functions.R
\name{anchor_fct}
\alias{anchor_fct}
\alias{anchor_extremal}
\alias{anchor_heuristic_extremal}
\title{Anchor functions for divisive cover}
\usage{
anchor_extremal(points, data, distance_fct)

anchor_heuristic_extremal(points, data, distance_fct)
}
\arguments{
\item{points}{indices of points in patch}

\item{data}{data}

\item{distance_fct}{distance function}
}
\description{
Anchor functions for divisive cover
}
