% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_functions.R
\name{anchor_fct}
\alias{anchor_fct}
\alias{anchor_extremal}
\alias{anchor_heuristic_extremal}
\alias{anchor_random_extremal}
\alias{anchor_classify}
\alias{anchor_heuristic_classify}
\title{Anchor functions for divisive cover}
\usage{
anchor_extremal(points, data, distance_fct, ...)

anchor_heuristic_extremal(points, data, distance_fct, ...)

anchor_random_extremal(points, data, distance_fct, ...)

anchor_classify(points, data, distance_fct, group, ...)

anchor_heuristic_classify(points, data, distance_fct, group, ...)
}
\arguments{
\item{points}{indices of points in patch}

\item{data}{data}

\item{distance_fct}{distance function}

\item{...}{ignored arguments}

\item{group}{the group for classification}
}
\description{
Find the anchor points used to divide a subset.
}
\details{
The function \code{anchor_extremal} returns the two points that are 
furthest apart. Use this function together with \code{\link{distance_matrix}}
as distance function. 

The function \code{anchor_heuristic_extremal} returns two points that 
approximate extremal points. Use this function together with
\code{\link{distance_cdist}} or \code{\link{distance_euclidean}}. 

The function \code{random_extremal} returns one random point and the point
that is furthest away from the randomly chosen point. 

The functions \code{anchor_classify} and \code{anchor_heuristic_classify} return 
one point from each of the two largest groups such that the two points are
extremal (or approximately extremal).
}
