% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R, R/cover_methods.R
\docType{class}
\name{cover-class}
\alias{cover-class}
\alias{show,cover-method}
\alias{show,patch-method}
\title{An S4 class to represent a cover.}
\usage{
\S4method{show}{cover}(object)

\S4method{show}{patch}(object)
}
\arguments{
\item{object}{A cover}
}
\description{
An S4 class to represent a cover.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame}

\item{\code{subsets}}{A list of patches}
}}

