% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{cover-class}
\alias{cover-class}
\title{An S4 class to represent a cover.}
\description{
An S4 class to represent a cover.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame}

\item{\code{subsets}}{A list of patches}
}}

