% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{patch-class}
\alias{patch-class}
\title{An S4 class to represent a patch.}
\description{
An S4 class to represent a patch.
}
\section{Slots}{

\describe{
\item{\code{indices}}{Indices of data points in patch}

\item{\code{predicted}}{Indices of predicted points}

\item{\code{children}}{Indices of children in cover}

\item{\code{parent}}{Indices of parent in cover}

\item{\code{birth}}{Birth time}

\item{\code{death}}{Death time}
}}

