% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{cover_coloring}
\alias{cover_coloring}
\alias{predict_coloring}
\title{Generate coloring for cover}
\usage{
cover_coloring(x, color_values, FUN = mean, ...)

predict_coloring(x)
}
\arguments{
\item{x}{cover}

\item{color_values}{value for each row}

\item{FUN}{summary function}

\item{...}{other orguments passed to FUN}
}
\description{
Generates a coloring for cover to be used with \code{\link{plot.cover}}
}
