% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_covers.R
\name{cover_distance}
\alias{cover_distance}
\alias{cover_distance_ph}
\alias{cover_distance_vi}
\title{Distance between covers}
\usage{
cover_distance_ph(cover1, cover2)

cover_distance_vi(cover1, cover2)
}
\arguments{
\item{cover1, cover2}{covers to compare}
}
\description{
Compare two covers using persistent homology or entropy
}
\examples{
# generate data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)
# calculate covers
dc1 <- divisive_cover(distance_matrix = dist(data_matrix), 
                      delta = 0.1,
                      relative_diameter = 0.1)
dc2 <- divisive_cover(distance_matrix = dist(data_matrix), 
                      delta = 0.05,
                      relative_diameter = 0.3)
cover1 <- subcover(dc1, 0.1, "snapshot")
cover2 <- subcover(dc2, 0.5, "snapshot")

cover_distance_ph(cover1, cover2)
cover_distance_vi(cover1, cover2)

# plot covers
\dontrun{
plot(cover1)
plot(cover2)
plot(c(cover1, cover2))
}
}

