% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{distance_fct}
\alias{distance_fct}
\alias{distance_cdist}
\alias{distance_euclidean}
\alias{distance_matrix}
\title{Distance functions for divisive cover}
\usage{
distance_cdist(metric = "euclidean")

distance_euclidean(data, X, Y)

distance_matrix(data, X, Y)
}
\arguments{
\item{metric}{metric in `cdist`}

\item{data}{data for divisive cover}

\item{X, Y}{indices}
}
\description{
Distance functions for divisive cover
}
