% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_ensemble.R
\name{divisive_classification_ensemble}
\alias{divisive_classification_ensemble}
\alias{predict_divisive_classification_ensemble}
\title{Divisive classification ensemble}
\usage{
divisive_classification_ensemble(train, group, ntree = 100, depth = 1000,
  delta_range = c(0, 0.2), anchor_fct = anchor_extremal,
  distance_fct = distance_cdist("euclidean"),
  stop_fct = stop_relative_filter_max_nodes(relative_filter = 0, max_nodes =
  depth), filter_fct = entropy_filter)

predict_divisive_classification_ensemble(dc_ensemble, test)
}
\arguments{
\item{train}{Training data}

\item{group}{Class of training data}

\item{ntree}{Number of trees to grow}

\item{depth}{Maximal depth of trees}

\item{delta_range}{Range of delta parameters to consider}

\item{anchor_fct}{Which anchor function should be used (see \code{\link{anchor_fct}})}

\item{distance_fct}{Which distance function should be used (see \code{\link{distance_fct}})}

\item{stop_fct}{Which stop function should be used (see \code{\link{stop_fct}})}

\item{filter_fct}{Which filter function should be used (see \code{\link{filter_fct}})}

\item{dc_ensemble}{Output of \code{divisive_classification_ensemble}}

\item{test}{Test data}
}
\description{
A divisive classification ensemble is an algorithm similar to random 
forests with divisive covers instead of decission trees. Currently uses
\code{\link{relative_gap_division}} with random gap as a division 
function.
}
\examples{
# data
sampled_index <- sample(150)
train <- iris[sampled_index[1:100], 1:4]
group <- iris[sampled_index[1:100], 5]
test <- iris[sampled_index[101:150], 1:4]
observed <- iris[sampled_index[101:150], 5]

# fit divisive classification ensemble
fit <- divisive_classification_ensemble(train = train, 
                                        group = group, 
                                        ntree = 2, 
                                        depth = 10, 
                                        delta_range = c(0, 0.1))
# get predictions
predictions <- predict_divisive_classification_ensemble(fit, test)

# confusion matrix
table(observed, predictions)
}
