% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover.R
\name{divisive_cover}
\alias{divisive_cover}
\title{divisive cover}
\usage{
divisive_cover(distance_matrix = NULL, relative_distance = 0.2,
  relative_diameter = 0.7)
}
\arguments{
\item{distance_matrix}{an n x n matrix of pairwise dissimilarities}

\item{relative_distance}{numeric to specify the amount of overlap in each division}

\item{relative_diameter}{numeric to specify the maximal diameter of the final cover}
}
\description{
This function divides the data into a cover using the distance matrix.
}
\examples{
rcircle <- function(N, r, sd){
   radius <- rnorm(N, r, sd)
   angle <- runif(N, 0, 2 * pi)
   data.frame(x = radius * cos(angle), 
              y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

dc <- divisive_cover(distance_matrix = dist(data_matrix), 
                     relative_diameter = 0.5, relative_distance = 0.2)
ddc <- diameter_cover(dc, relative_diameter = 0.7)

\dontrun{
plot(data_matrix)
plot(ddc)
}
}

