% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover.R
\name{divisive_cover}
\alias{divisive_cover}
\title{divisive cover}
\usage{
divisive_cover(cover = NULL, data,
  distance_fct = distance_cdist("euclidean"),
  stop_fct = stop_relative_filter(relative_filter = 0.5),
  anchor_fct = anchor_extremal, filter_fct = diameter_filter,
  division_fct = relative_gap_division(relative_gap = 0.1))
}
\arguments{
\item{cover}{a divisive cover used to update}

\item{stop_fct}{function that determines when to stop dividing (see details below)}

\item{distance_matrix}{an n x n matrix of pairwise dissimilarities}
}
\description{
This function divides the data into a cover using the distance matrix.
}
\details{
The function \code{stop_fct} is a function of ... that returns \code{TRUE} if the division should
be stoped and \code{FALSE} otherwise. Examples are \code{\link{stop_relative_filter}} and 
\code{\link{stop_max_nodes}}.
}
\examples{
# generate sample data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# calculate and update divisive cover
dc1 <- divisive_cover(distance_matrix = dist(data_matrix),
                      delta = 0.05, 
                      stop_fct = stop_relative_filter(0.5))
dc2 <- divisive_cover(cover = dc1, 
                      distance_matrix = dist(data_matrix),
                      delta = 0.05, 
                      stop_fct = stop_max_nodes(20))

# get one snapshot for plotting
ddc1 <- subcover(dc1, relative_diameter = 0, method = "snapshot")
ddc2 <- subcover(dc2, relative_diameter = 0, method = "snapshot")

\dontrun{
plot(data_matrix)
plot(ddc1)
plot(ddc2)
}
}

