% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover.R
\name{divisive_cover}
\alias{divisive_cover}
\title{divisive cover}
\usage{
divisive_cover(data_matrix = NULL, distance_object = NULL,
  distance_mode = guess_distance_mode(distance_object),
  relative_distance = 0.1, relative_radius = 0.7, base_point = NULL)
}
\arguments{
\item{data_matrix}{An data.frame or matrix (required if distance_object is a function)}

\item{distance_object}{an n x n matrix of pairwise dissimilarities or a distance function 
such as \code{dist} or \code{flexclust::dist2}}

\item{distance_mode}{If \code{distance_object} is a matrix, \code{distance_mode} should be 
\code{"matrix"}, if \code{distance_object} is a function that takes 1 argument as a 
input and returns pairwise distances, then  \code{distance_mode} should be \code{"dist"}, 
and if \code{distance_object} is a function that takes 2 matrices as input and returns 
the distances between these two matrices, then \code{distance_mode} should be \code{"dist2"}.}

\item{gap}{numeric to specify the amount of overlap in each division}

\item{numeric}{to specify the maximal diameter of the final cover}
}
\description{
This function divides the data into a cover using the distance matrix.
}

