% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover.R
\name{divisive_cover}
\alias{divisive_cover}
\title{divisive cover}
\usage{
divisive_cover(cover = NULL, data,
  distance_fct = distance_cdist("euclidean"),
  stop_fct = stop_relative_filter(relative_filter = 0.5),
  anchor_fct = anchor_extremal, filter_fct = diameter_filter,
  division_fct = relative_gap_division(relative_gap = 0.1), group = NULL)
}
\arguments{
\item{cover}{an optional divisive cover used to update}

\item{data}{a data matrix}

\item{distance_fct}{the distance function used (see details below)}

\item{stop_fct}{function that determines when to stop dividing (see details below)}

\item{anchor_fct}{function that determined how anchor points are found (see details below)}

\item{filter_fct}{function that filters the cover (see details below)}

\item{division_fct}{function that divides the patches (see details below)}

\item{group}{group for classification}
}
\description{
This function divides the data into a cover using the distance matrix.
}
\details{
The function \code{distance_function} is a function of \code{data}, \code{X} and \code{Y}, where
\code{data} is the data matrix and \code{X} and \code{Y} are indices. Examples of distance functions
are \code{\link{distance_euclidean}}, \code{\link{distance_matrix}}, and the function returned from 
\code{\link{distance_cdist}}. 

The function \code{stop_fct} is a function of \code{cover} and \code{next_division}, where \code{next_division} 
is the index of the next subset that should be divided. The function returns
\code{TRUE} if the division should be stoped and \code{FALSE} otherwise. Examples are returned from
\code{\link{stop_relative_filter}}, \code{\link{stop_max_nodes}}, and \code{\link{stop_relative_filter_max_nodes}}.

The function \code{anchor_fct} is a function of \code{points}, the indices of points in patch, 
\code{data}, \code{distance_fct}, and \code{group}. Examples are \code{\link{anchor_extremal}}, 
\code{\link{anchor_heuristic_extremal}}, \code{\link{anchor_classify}}, and \code{\link{anchor_heuristic_classify}}

TODO: Document filter_fct, division_fct
}
\examples{
# generate sample data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# calculate and update divisive cover
dc1 <- divisive_cover(data = data_matrix,
                      division_fct = relative_gap_division(0.05), 
                      stop_fct = stop_relative_filter(0.5))
dc2 <- divisive_cover(cover = dc1, 
                      data = data_matrix,
                      division_fct = relative_gap_division(0.05),
                      stop_fct = stop_max_nodes(20))

# get one snapshot for plotting
ddc1 <- subcover(dc1, relative_filter = 0, method = "snapshot")
ddc2 <- subcover(dc2, relative_filter = 0, method = "snapshot")

\dontrun{
plot(data_matrix)
plot(ddc1)
plot(ddc2)
}
}
