% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_cover.R
\name{divisive_cover}
\alias{divisive_cover}
\title{divisive cover}
\usage{
divisive_cover(cover = NULL, data,
  distance_fct = distance_cdist("euclidean"),
  stop_fct = stop_relative_filter(relative_filter = 0.5),
  anchor_fct = anchor_extremal, filter_fct = diameter_filter,
  division_fct = relative_gap_division(relative_gap = 0.1))
}
\arguments{
\item{cover}{an optional divisive cover used to update}

\item{data}{a data matrix}

\item{distance_fct}{the distance function used (see details below)}

\item{stop_fct}{function that determines when to stop dividing (see details below)}

\item{anchor_fct}{function that determined how anchor points are found (see details below)}

\item{filter_fct}{function that filters the cover (see details below)}

\item{division_fct}{function that divides the patches (see details below)}
}
\description{
This function divides the data into a cover using the distance matrix.
}
\details{
TODO... write detailed documentation
The function \code{stop_fct} is a function of ... that returns \code{TRUE} if the division should
be stoped and \code{FALSE} otherwise. Examples are \code{\link{stop_relative_filter}} and 
\code{\link{stop_max_nodes}}.
}
\examples{
# generate sample data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# calculate and update divisive cover
dc1 <- divisive_cover(data = data_matrix,
                      division_fct = relative_gap_division(0.05), 
                      stop_fct = stop_relative_filter(0.5))
dc2 <- divisive_cover(cover = dc1, 
                      data = data_matrix,
                      division_fct = relative_gap_division(0.05),
                      stop_fct = stop_max_nodes(20))

# get one snapshot for plotting
ddc1 <- subcover(dc1, relative_filter = 0, method = "snapshot")
ddc2 <- subcover(dc2, relative_filter = 0, method = "snapshot")

\dontrun{
plot(data_matrix)
plot(ddc1)
plot(ddc2)
}
}

