% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cover.R
\name{extract_cover}
\alias{data_from_cover_id}
\alias{extract_cover}
\alias{index_from_cover_id}
\title{Extract data from cover}
\usage{
index_from_cover_id(cover, ids)

data_from_cover_id(data, cover, ids)
}
\arguments{
\item{cover}{A cover}

\item{ids}{Cover ids, for example extracted from plot}

\item{data}{Data from which to extract subset}
}
\description{
Functions to extract relevant data from covers
}

