% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_fct}
\alias{filter_fct}
\alias{diameter_filter}
\title{Filter functions}
\usage{
diameter_filter(patches, data, distance_fct)
}
\arguments{
\item{patches}{Patch to apply filter}

\item{data}{Data used in distance function}

\item{distance_fct}{The used distance function}
}
\description{
Filter after diameter or cardinality
}
