% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_fct}
\alias{filter_fct}
\alias{diameter_filter}
\alias{cardinality_filter}
\alias{classification_filter}
\alias{entropy_filter}
\title{Filter functions}
\usage{
diameter_filter(patches, data, distance_fct, ...)

cardinality_filter(patches, ...)

classification_filter(patches, group, ...)

entropy_filter(patches, group, ...)
}
\arguments{
\item{patches}{Patch to apply filter}

\item{data}{Data used in distance function}

\item{distance_fct}{The used distance function}

\item{...}{ignored arguments}

\item{group}{the group for classification}
}
\description{
Filter the divisive cover. The filter is used to decide which subset 
to divide next and may be used to decide when to stop dividing. It is 
also used for calculating persistent homology.
}
\details{
The function \code{diameter_filter} returns the distance between the
two anchor points. 

The function \code{cardinality_filter} returns the cardinality of the
subset. 

The function \code{classification_filter} returns the proportion of 
elements in a subset that do not belong to the largest group.
}
