% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{plot.cover}
\alias{plot.cover}
\title{Plot cover output}
\usage{
\method{plot}{cover}(x, coloring = NULL, simplify = c("none", "duplicates",
  "subsets"), label = sapply(x@subsets, slot, "id"), legend = TRUE,
  device = c("plot", "tkplot"), seed = 1, ...)
}
\arguments{
\item{x}{A cover}

\item{coloring}{color values. See \code{\link{cover_coloring}}}

\item{simplify}{simplification method. See \code{\link{simplify_cover}}}

\item{label}{how should the nodes be labeled. Set NA for no labels.}

\item{legend}{should color legend be displayed}

\item{device}{use igraph::plot.igraph or igraph::tkplot}

\item{seed}{random seed used to make plots look the same}

\item{...}{additional plotting parameters. See \code{\link[igraph]{igraph.plotting}} for the complete list.}
}
\description{
\code{plot.cover} generates R plots of covers
}
\seealso{
\code{\link{cover-class}}, \code{\link{subcover}}
}
\author{
Nello Blaser
}
\keyword{plotting}
