% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cover}
\alias{predict.cover}
\title{Predict}
\usage{
\method{predict}{cover}(object, newdata, predict_fct, ...)
}
\arguments{
\item{object}{a divisive cover}

\item{newdata}{new data}

\item{predict_fct}{predict function (see below)}

\item{...}{further arguments passed from other methods}
}
\description{
Predict in which subsets new data come to lie when given a previous
divisive cover
}
\examples{
# generate data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# run divisive cover
dc <- divisive_cover(data = data_matrix, 
                     division_fct = relative_gap_division(0.1), 
                     stop_fct = stop_relative_filter(relative_filter = 0.2))
                     
# predict
pred <- predict(dc, newdata = rcircle(10, 1, .1), 
                predict_fct = relative_gap_prediction(0.1))

# plot
sc <- subcover(pred, 0.6, "snapshot")
col <- predict_coloring(sc)
\dontrun{
plot(sc, coloring = col)
}
}
