% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cover}
\alias{predict.cover}
\title{Predict}
\usage{
\method{predict}{cover}(object, newdists = NULL, newdata = NULL, ...)
}
\arguments{
\item{object}{a divisive cover}

\item{newdists}{a matrix of distances between the new points and the 
points used to train the divisive cover}

\item{newdata}{new data in case of fast divisive cover}

\item{...}{further arguments passed from other methods}
}
\description{
Predict in which subsets new data come to lie when given a previous
divisive cover
}
\examples{
# generate data
rcircle <- function(N, r, sd){
  radius <- rnorm(N, r, sd)
  angle <- runif(N, 0, 2 * pi)
  data.frame(x = radius * cos(angle), 
             y = radius * sin(angle))
}
data_matrix <- rcircle(200, 1, .1)

# run divisive cover
dc <- divisive_cover(distance_matrix = dist(data_matrix), 
                     delta = 0.1, 
                     stop_fct = stop_relative_diameter(relative_diameter = 0.2))
                     
# predict
pred <- predict(dc, newdists = as.matrix(dist(data_matrix))[, 1:10])

# plot
sc <- subcover(pred, 0.6, "snapshot")
col <- predict_coloring(sc)
\dontrun{
plot(sc, coloring = col)
}
}

