% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{stop_fct}
\alias{stop_fct}
\alias{stop_max_nodes}
\alias{stop_relative_diameter}
\title{Stop Functions}
\usage{
stop_relative_diameter(relative_diameter)

stop_max_nodes(max_nodes)
}
\arguments{
\item{relative_diameter}{maximal diameter of the final cover}

\item{max_nodes}{maximum number of nodes the algorithm should generate}
}
\description{
Possible stop criteria are based on the relative diameter 
or the maximum number of nodes.
}

