% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{stop_fct}
\alias{stop_fct}
\alias{stop_relative_filter}
\alias{stop_max_nodes}
\alias{stop_relative_filter_max_nodes}
\title{Stop Functions}
\usage{
stop_relative_filter(relative_filter)

stop_max_nodes(max_nodes)

stop_relative_filter_max_nodes(relative_filter, max_nodes)
}
\arguments{
\item{relative_filter}{maximal relative filter value of final cover}

\item{max_nodes}{maximum number of nodes the algorithm should generate}
}
\description{
Possible stop criteria are based on the relative filter value 
or the maximum number of nodes.
}
\details{
These functions return stop functions that can be used in \code{\link{divisive_cover}}. The stop functions
return \code{TRUE} if the division should be stoped and \code{FALSE} otherwise.

The function \code{stop_relative_filter} returns a function that stops the division when the subset that should 
next be divided has a relative filter value of less than \code{relative_filter}. 

The function \code{stop_max_nodes} returns a function that stops the division when \code{max_nodes} nodes have 
been created. 

The function \code{stop_relative_filter_max_nodes} returns a function which stops the division if one of the 
above criteria is satisfied.
}
